Refactor tree structure tests
Each node type has its own file describing the possible parents it can
have along with its list of children. This is to allow for the
documentation to easily include the actual values that are tested.
