+++ tests/tree_structure/Literal.py
@@ -1,0 +1,5 @@
+from tests.tree_structure import literal, predicate, either, zero_or_more, maybe
+from parse_ebnf.nodes import *
+
+parents = [ Comment, Product, DefinitionList, Definition, Term, Repetition, Terminal, Repeat, Option, Group, Special ]
+children = [ ]
