+++ tests/tree_structure/Special.py
@@ -1,0 +1,5 @@
+from tests.tree_structure import literal, predicate, either, zero_or_more, maybe
+from parse_ebnf.nodes import *
+
+parents = [ Term, Exception ]
+children = [ literal("?"), Text, literal("?") ]
